+++ src/Application/Features/Trading/SellStock/SellStockCommand.cs
@@ -3,5 +3,5 @@ using OsuStocks.Application.Common.Models;
 
 namespace OsuStocks.Application.Features.Trading.SellStock;
 
-public sealed record SellStockCommand(Guid UserId, Guid StockId, int Quantity)
+public sealed record SellStockCommand(Guid UserId, Guid StockId, decimal Quantity)
     : IRequest<Result<SellStockResponse>>;
