+++ src/Application/Features/Trading/Services/ITradingGuardService.cs
@@ -12,8 +12,8 @@ public interface ITradingGuardService
     Task<Result> CheckPositionLimitAsync(
         Guid userId,
         Guid stockId,
-        int requestedQuantity,
-        int currentHoldingQuantity,
+        decimal requestedQuantity,
+        decimal currentHoldingQuantity,
         CancellationToken cancellationToken = default);
 
     Task CheckRapidTradingAsync(
